variants: Rename variants.h to variant.h
The header for Arduino board-specific settings is variant.h,
not variants.h, so I modified it accordingly.
