+++ documentation/variants.md
@@ -27,7 +27,7 @@ target board. To add board support:
 needs to be added it needs to be done in the `variants/` folder.
 Add a folder inside of the variants folder that matches the name of your board.
 2. Add an overlay file and a pinmap header file that match the name of the board.
-3. Add your new headerfile to an `#ifdef` statement in the variants.h file.
+3. Add your new headerfile to an `#ifdef` statement in the variant.h file.
 
 An example of this structure is shown below.
 
@@ -37,13 +37,13 @@ variants/
 │   ├── arduino_nano_33_ble.overlay
 │   └── arduino_nano_33_ble_pinmap.h
 ├── CMakeLists.txt
-└── variants.h
+└── variant.h
 
 ```
 
-- The top level consists of `CMakeLists.txt`, `variants.h` and the `<BOARD_NAME>` folder. Each of these files have a specific role to play.
+- The top level consists of `CMakeLists.txt`, `variant.h` and the `<BOARD_NAME>` folder. Each of these files have a specific role to play.
 		- The `Cmakelists` help the compiler locate the proper directory to help find the proper header files that are board specific. You need to add the name using `zephyr_include_directories(BOARD_NAME)` to this file. Do note that this `BOARD_NAME` is the same as the name of your board's directory.
-		- `variants.h` contains the necessary `#includes` inorder to tell the source code about your board's pinmap.
+		- `variant.h` contains the necessary `#includes` inorder to tell the source code about your board's pinmap.
 - The `<BOARD_NAME>` folder is where the overlay and pinmap file resides. Inorder to understand how to write DT overlays, lookup `Documentation/overlays.md`. To understand the `<boardname_pinmap.h>` file, go through the existing `variants/ARDUINO_NANO33BLE/arduino_nano_ble_sense_pinmap.h` which shows how to use the overlay nodes inside our C programs using zephyr macros like `GPIO_DT_SPEC_GET`. The zephyr-project documentation on this is pretty extensive as well and worth reading.
 
 ## Guide to Writing Overlays
@@ -193,7 +193,7 @@ One example of a change that you may find useful is mapping additional pins. For
 example, the LEDs on the nRF52840 are not connected to any of the Arduino header
 pins. To define a built-in LED for this board, a 22nd pin definition was added.
 
-Your pinmap header file must be added to the variants.h file by adding three
+Your pinmap header file must be added to the variant.h file by adding three
 lines using this format:
 
 ```c
