+++ variants/beagleconnect_freedom/variant.h
