+++ variants/cc3220sf_launchxl/variant.h
