+++ variants/nrf52840dk_nrf52840/variant.h
