+++ variants/nrf9160dk_nrf9160/variant.h
